BLD: Add pyproject.toml (#241)
* BLD: Initial try at pyproject.toml

I think I got every key from setup.py.
The main difference is if the READTHEDOCS environment variable is set: I don't know how to change the dependencies based on that.

On the other hand, given 3.8 is close to or past EOL, I suspect 3.3 is relatively uncommon, and the RTD install, without the unneeded bits, could be accomplished with pip install --no-deps.

* BLD,BUG: Fix pyproject.toml configuration and remove setup.py duplicates.

numpy.distutils recommends setuptools<60, but that breaks pyproject.toml
configuration.  I like this way better.  Still needs setup.py for the
fortran module and for the Cheyenne configuration.

* FIX: Require a setuptools version that understands the configuration.

setuptools<60 just gets confused, and has no idea what the project name is.
I just inverted that requirement: not sure if setuptools needs to be more recent still.

* BLD: Quote version requirement.

* BLD: Specify dependencies in pyproject.toml only.

Remove requirements.txt.

* DOC: Don't install dependencies on RTD

The old setup script replaced the dependencies listed in requirements.txt with mock if python was old enough and nothing if python was recent.

The package no longer supports python versions that old (<3.3), so installing without dependencies is likely the correct way to handle the RTD configuration.

* BLD,BUG: Remove references to requirements, not just definitions

Forgot to remove this two or three commits back.

* BLD: Let's see if NumPy < 2.0 fixes the error

* BLD: Pin setuptools<60

NumPy suggested it, it might help with the distutils failure.

* BLD: Pin setuptools<71

Let's see if that change is the one that broke things.

* BLD: Add subroutines to f90 modules so f2py adds code for them.

If it's just constants used somewhere, f2py drops the module.

* BLD: Remove build-time setuptools pin.

* BLD: Pin setuptools<70 to get msvccompiler

* BLD: Add -fopenmp to LDFLAGS

Let's see if this gets Linux compiles working.

---------

Co-authored-by: Katelyn FitzGerald <7872563+kafitzgerald@users.noreply.github.com>